Add a new "Product Packaging" field on quant package
Instead of reusing the existing 'packaging_id'.

Odoo is using the "product.packaging" model for 2 different purposes:

1. Records without 'product_id' are packaging types, used for the
delivery, such as the packages types of the postal office. This is
actually the use case for the existing field 'package_id' on
'stock.quant.package' (there is a domain ("product_id", "=",
False)).
2. Records with a 'product_id' define the logistics packaging types
of the product (a pallet of 500, that contains 50 boxes of 10).

Since the first field has a domain on the first type of packaging,
the "product packaging" is now stored in a second domain.

The module stock_quant_package_product_packaging_delivery is no longer
needed since there is no conflict with the delivery module now.